fairly handled bottlenecks like different type of file modifications in one object and storing each chunk first in disk then getting every chunk as sorted, writing to disk is efficient for large files!

+++ src/file-pair-session.cpp
@@ -41,35 +41,54 @@ void FilePairSession::reset_if_filepath_changes_append_required(const std::strin
     ensure_files_open();
 }
 
-// TODO: optimise in case of filling large gap
 // Fills gap in temp_file if present from the offset
 void FilePairSession::fill_gap_till_offset(const size_t offset)
 {
-    // skip if there is no gap to be filled
-    if (offset <= cursor)
+    // skip if there is no gap to be filled or temp_file not exists
+    if (offset <= cursor || !temp_file)
         return;
 
-    // data copied to be filled
-    std::string to_copy = original_file->read_file_from_offset(cursor, offset);
+    size_t chunk_size = offset - cursor;
+
+    for (size_t read_so_far = 0; read_so_far < chunk_size;)
+    {
+        const std::string &to_copy = original_file->read_file_from_offset(
+            cursor + read_so_far,
+            std::min(MAX_READ_SIZE, chunk_size - read_so_far));
+
+        if (to_copy.empty())
+            break;
+
+        temp_file->append_chunk(to_copy);
+
+        read_so_far += to_copy.size();
+    }
 
     // as we have moved BYTES forward after copying BYTES data
     cursor = offset;
-
-    // now fill the data to temp_file
-    temp_file->append_chunk(to_copy);
 }
 
 // append chunk data to original file
 void FilePairSession::append_data(const std::string &chunk)
 {
-    original_file->append_chunk(chunk);
+    if (is_appending_to_original())
+        original_file->append_chunk(chunk);
 }
 
 // chunk_size can be != chunk.size as in case of modify we will be replacing data of differet size
-void FilePairSession::add_chunk(const std::string &chunk, const size_t chunk_size)
+void FilePairSession::add_chunk(const std::string &chunk, const size_t chunk_size, const bool is_new_chunk)
 {
+    if (!temp_file)
+    {
+        std::cerr << "failed to add chunk! temp_file not exists" << std::endl;
+        return;
+    }
+
     temp_file->append_chunk(chunk);
-    cursor += chunk_size;
+
+    // since the chunk was not previously present so no need to move the original file cursor
+    if (!is_new_chunk)
+        cursor += chunk_size;
 }
 
 // skip the given chunk from writing to temp_file
@@ -78,14 +97,16 @@ void FilePairSession::skip_removed_chunk(const size_t offset, const size_t chunk
     cursor = offset + chunk_size;
 }
 
-// TODO: optimise in case of filling large remaining data
 // fill rest of the data from original file to temp_file + rename temp_file to original
 void FilePairSession::finalize_and_replace()
 {
-    fill_gap_till_offset(original_file->get_file_size() - 1);
+    if (temp_file)
+    {
+        fill_gap_till_offset(original_file->get_file_size());
 
-    fs::remove(original_filepath);
-    fs::rename(temp_filepath, original_filepath);
+        fs::remove(original_filepath);
+        fs::rename(temp_filepath, original_filepath);
+    }
 
     close_session();
 }
@@ -113,4 +134,9 @@ std::string FilePairSession::get_filepath()
 bool FilePairSession::is_appending_to_original()
 {
     return append_to_original;
+}
+
+FilePairSession::~FilePairSession()
+{
+    close_session();
 }
